# openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/approval-workflows/{id}" : {
      "get" : {
        "tags" : [ "approval-workflow-definition-controller" ],
        "operationId" : "get_9",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DefinitionView"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "approval-workflow-definition-controller" ],
        "operationId" : "update_6",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpsertRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DefinitionView"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "approval-workflow-definition-controller" ],
        "operationId" : "delete_6",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
...
    "/api/v1/approval-workflows" : {
      "get" : {
        "tags" : [ "approval-workflow-definition-controller" ],
        "operationId" : "list_8",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/DefinitionView"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "approval-workflow-definition-controller" ],
        "operationId" : "create_11",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpsertRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DefinitionView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/users/{id}/identities" : {
      "get" : {
        "tags" : [ "user-identity-controller" ],
        "operationId" : "list_10",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/IdentityView"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "user-identity-controller" ],
        "operationId" : "link",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LinkRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IdentityView"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "user-identity-controller" ],
        "operationId" : "unlink",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "issuer",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "subject",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
...
    "/api/v1/admin/users/identities/import" : {
      "post" : {
        "tags" : [ "user-identity-controller" ],
        "operationId" : "importIdentities",
        "parameters" : [ {
          "name" : "dryRun",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : true
          }
        } ],
        "requestBody" : {
          "content" : {
            "text/plain" : {
              "schema" : {
                "type" : "string"
              }
            },
            "application/json" : {
              "schema" : {
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/approvals/{requestId}" : {
      "get" : {
        "tags" : [ "approval-controller" ],
        "operationId" : "get_18",
        "parameters" : [ {
          "name" : "requestId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApprovalDetailView"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "StepInput" : {
        "type" : "object",
        "properties" : {
          "stepOrder" : {
            "type" : "integer",
            "format" : "int32"
          },
          "slaHours" : {
            "type" : "integer",
            "format" : "int32"
          },
          "approverRoleId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "escalatesToStepOrder" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "stepOrder" ]
      },
      "UpsertRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "minLength" : 1
          },
          "scopeSpaceId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "scopeDocType" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "minLength" : 1
          },
          "steps" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/StepInput"
            },
            "minItems" : 1
          }
        },
        "required" : [ "name", "status", "steps" ]
      },
      "DefinitionView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "scopeSpaceId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "scopeDocType" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string"
          },
          "steps" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/StepView"
            }
          },
          "inProgressCount" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "StepView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "stepOrder" : {
            "type" : "integer",
            "format" : "int32"
          },
          "slaHours" : {
            "type" : "integer",
            "format" : "int32"
          },
          "approverRoleId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "approverRoleName" : {
            "type" : "string"
          },
          "escalatesToStepOrder" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "LinkRequest" : {
        "type" : "object",
        "properties" : {
          "issuer" : {
            "type" : "string",
            "minLength" : 1
          },
          "subject" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "issuer", "subject" ]
      },
      "IdentityView" : {
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "issuer" : {
            "type" : "string"
          },
          "subject" : {
            "type" : "string"
          },
          "linkedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "linkedBy" : {
            "type" : "string",
            "format" : "uuid"
          }
        }
      },
...
      "ImpactedLink" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "title" : {
            "type" : "string"
          }
        }
      },
...
      "ApprovalDetailView" : {
        "type" : "object",
        "properties" : {
          "approvalRequestId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "documentId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "documentTitle" : {
            "type" : "string"
          },
          "temporalWorkflowId" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "currentStepOrder" : {
            "type" : "integer",
            "format" : "int32"
          },
          "slaDeadlineAt" : {
            "type" : "string"
          },
          "submittedVersionNo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "baselineVersionNo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "requestedBy" : {
            "type" : "string",
            "format" : "uuid"
          },
          "createdAt" : {
            "type" : "string"
          },
          "requestedByDisplayName" : {
            "type" : "string"
          },
          "requestedByInitials" : {
            "type" : "string"
          },
          "impactedLinks" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ImpactedLink"
            }
          },
          "hiddenImpactedCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "canDecide" : {
            "type" : "boolean"
          },
          "cannotDecideReason" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}